+++ Common/Budget/IBudgetLedger.cs
@@ -12,6 +12,6 @@ namespace Lucent.Common.Budget
         /// <summary>
         /// Tries to record an entry with the source and amount
         /// </summary>
-        Task<bool> TryRecordEntry<T>(string ledgerId, T source, EntityType eType, decimal amount) where T : class, new();
+        Task<bool> TryRecordEntry<T>(string ledgerId, T source, EntityType eType, double amount) where T : class, new();
     }
 }
